Main/Appliance: MSVC compiler warnings
svn:sync-xref-src-repo-rev: r67332

+++ src/VBox/Main/ApplianceImplIO.cpp
@@ -1,4 +1,4 @@
-/* $Id: ApplianceImplIO.cpp 33659 2010-11-01 14:46:15Z noreply@oracle.com $ */
+/* $Id: ApplianceImplIO.cpp 33706 2010-11-02 18:41:56Z klaus.espenlaub@oracle.com $ */
 /** @file
  *
  * IO helper for IAppliance COM class implementations.
@@ -955,7 +955,7 @@ static int rtSha1WriteSyncCallback(void *pvUser, void *pvStorage, uint64_t uOffs
      * real write. */
     if (pInt->cbCurAll < uOffset)
     {
-        size_t cbSize = uOffset - pInt->cbCurAll;
+        size_t cbSize = (size_t)(uOffset - pInt->cbCurAll);
         size_t cbAllWritten = 0;
         for(;;)
         {
@@ -1052,7 +1052,8 @@ static int rtSha1ReadSyncCallback(void *pvUser, void *pvStorage, uint64_t uOffse
      * remaining stuff in the gap anyway (SHA1; streaming). */
     if (pInt->cbCurAll < uOffset)
     {
-        rc = rtSha1ReadSyncCallback(pvUser, pvStorage, pInt->cbCurAll, 0, uOffset - pInt->cbCurAll, 0);
+        rc = rtSha1ReadSyncCallback(pvUser, pvStorage, pInt->cbCurAll, 0,
+                                    (size_t)(uOffset - pInt->cbCurAll), 0);
         if (RT_FAILURE(rc))
             return rc;
     }
